New platform move doc table and doc viewer to kibana\discover (#34231)
* Moved doc table to kibana/discover

* moved doc viewer to Kibana app (used by docs and discover)

* Moved
- saved object finder
- paginated selectable list
from partiasls to directives/partials

* Moved paginate controls directive to directives/partials
