+++ src/legacy/core_plugins/kibana/public/discover/_index.scss
@@ -10,6 +10,7 @@
 @import 'components/fetch_error/index';
 @import 'components/field_chooser/index';
 @import 'directives/index';
+@import 'doc_table/index';
 
 @import 'hacks';
