+++ src/legacy/core_plugins/kibana/public/discover/doc_table/__tests__/actions/filter.js
@@ -18,7 +18,7 @@
  */
 
 import { addFilter } from '../../actions/filter';
-import { FilterManagerProvider } from '../../../filter_manager';
+import { FilterManagerProvider } from 'ui/filter_manager';
 import StubbedLogstashIndexPatternProvider from 'fixtures/stubbed_logstash_index_pattern';
 import NoDigestPromises from 'test_utils/no_digest_promises';
 import expect from '@kbn/expect';
