+++ src/legacy/core_plugins/kibana/public/discover/doc_table/__tests__/lib/get_sort.js
