+++ src/legacy/core_plugins/kibana/public/discover/doc_table/__tests__/lib/rows_headers.js
