+++ src/legacy/core_plugins/kibana/public/discover/doc_table/_doc_table.scss
