+++ src/legacy/core_plugins/kibana/public/discover/doc_table/_index.scss
