+++ src/legacy/core_plugins/kibana/public/discover/doc_table/components/_index.scss
