+++ src/legacy/core_plugins/kibana/public/discover/doc_table/components/_table_header.scss
