+++ src/legacy/core_plugins/kibana/public/discover/doc_table/components/table_header.html
@@ -7,7 +7,7 @@
   >
     <span>
       <span
-        i18n-id="common.ui.docTable.tableHeader.timeHeaderCellTitle"
+        i18n-id="kbn.docTable.tableHeader.timeHeaderCellTitle"
         i18n-default-message="Time"
       ></span>
       <button
@@ -17,7 +17,7 @@
       ng-class="headerClass(indexPattern.timeFieldName)"
       role="button"
       ng-click="cycleSortOrder(indexPattern.timeFieldName)"
-      tooltip="{{ ::'common.ui.docTable.tableHeader.sortByTimeTooltip' | i18n: {defaultMessage: 'Sort by time'} }}"
+      tooltip="{{ ::'kbn.docTable.tableHeader.sortByTimeTooltip' | i18n: {defaultMessage: 'Sort by time'} }}"
     ></button>
     </span>
   </th>
@@ -46,8 +46,8 @@
       ng-click="onRemoveColumn(name)"
       ng-if="canRemoveColumn(name)"
       tooltip-append-to-body="1"
-      tooltip="{{ ::'common.ui.docTable.tableHeader.removeColumnButtonTooltip' | i18n: {defaultMessage: 'Remove column'} }}"
-      aria-label="{{ 'common.ui.docTable.tableHeader.removeColumnButtonAriaLabel' | i18n: {
+      tooltip="{{ ::'kbn.docTable.tableHeader.removeColumnButtonTooltip' | i18n: {defaultMessage: 'Remove column'} }}"
+      aria-label="{{ 'kbn.docTable.tableHeader.removeColumnButtonAriaLabel' | i18n: {
         defaultMessage: 'Remove {columnName} column',
         values: {columnName: name}
       } }}"
@@ -58,8 +58,8 @@
       ng-click="moveColumnLeft(name)"
       ng-if="canMoveColumnLeft(name)"
       tooltip-append-to-body="1"
-      tooltip="{{ ::'common.ui.docTable.tableHeader.moveColumnLeftButtonTooltip' | i18n: {defaultMessage: 'Move column to the left'} }}"
-      aria-label="{{ 'common.ui.docTable.tableHeader.moveColumnLeftButtonAriaLabel' | i18n: {
+      tooltip="{{ ::'kbn.docTable.tableHeader.moveColumnLeftButtonTooltip' | i18n: {defaultMessage: 'Move column to the left'} }}"
+      aria-label="{{ 'kbn.docTable.tableHeader.moveColumnLeftButtonAriaLabel' | i18n: {
         defaultMessage: 'Move {columnName} column to the left',
         values: {columnName: name}
       } }}"
@@ -69,8 +69,8 @@
       ng-click="moveColumnRight(name)"
       ng-if="canMoveColumnRight(name)"
       tooltip-append-to-body="1"
-      tooltip="{{ ::'common.ui.docTable.tableHeader.moveColumnRightButtonTooltip' | i18n: {defaultMessage: 'Move column to the right'} }}"
-      aria-label="{{ 'common.ui.docTable.tableHeader.moveColumnRightButtonAriaLabel' | i18n: {
+      tooltip="{{ ::'kbn.docTable.tableHeader.moveColumnRightButtonTooltip' | i18n: {defaultMessage: 'Move column to the right'} }}"
+      aria-label="{{ 'kbn.docTable.tableHeader.moveColumnRightButtonAriaLabel' | i18n: {
         defaultMessage: 'Move {columnName} column to the right',
         values: {columnName: name}
       } }}"
