+++ src/legacy/core_plugins/kibana/public/discover/doc_table/components/table_row/_details.scss
