+++ src/legacy/core_plugins/kibana/public/discover/doc_table/components/table_row/cell.html
@@ -17,16 +17,16 @@ if (timefield) {
         class="fa fa-search-plus kbnDocTableRowFilterButton"
         data-column="<%- column %>"
         tooltip-append-to-body="1"
-        tooltip="{{ ::'common.ui.docTable.tableRow.filterForValueButtonTooltip' | i18n: {defaultMessage: 'Filter for value'} }}"
-        aria-label="{{ ::'common.ui.docTable.tableRow.filterForValueButtonAriaLabel' | i18n: {defaultMessage: 'Filter for value'} }}"
+        tooltip="{{ ::'kbn.docTable.tableRow.filterForValueButtonTooltip' | i18n: {defaultMessage: 'Filter for value'} }}"
+        aria-label="{{ ::'kbn.docTable.tableRow.filterForValueButtonAriaLabel' | i18n: {defaultMessage: 'Filter for value'} }}"
       ></button>
 
       <button
         ng-click="inlineFilter($event, '-')"
         class="fa fa-search-minus kbnDocTableRowFilterButton"
         data-column="<%- column %>"
-        tooltip="{{ ::'common.ui.docTable.tableRow.filterOutValueButtonTooltip' | i18n: {defaultMessage: 'Filter out value'} }}"
-        aria-label="{{ ::'common.ui.docTable.tableRow.filterOutValueButtonAriaLabel' | i18n: {defaultMessage: 'Filter out value'} }}"
+        tooltip="{{ ::'kbn.docTable.tableRow.filterOutValueButtonTooltip' | i18n: {defaultMessage: 'Filter out value'} }}"
+        aria-label="{{ ::'kbn.docTable.tableRow.filterOutValueButtonAriaLabel' | i18n: {defaultMessage: 'Filter out value'} }}"
         tooltip-append-to-body="1"
       ></button>
     <% } %>
