+++ src/legacy/core_plugins/kibana/public/discover/doc_table/components/table_row/open.html
@@ -2,7 +2,7 @@
   <button
     class="euiButtonIcon euiButtonIcon--text"
     aria-expanded="{{!!open}}"
-    aria-label="{{ ::'common.ui.docTable.tableRow.toggleRowDetailsButtonAriaLabel' | i18n: {defaultMessage: 'Toggle row details'} }}"
+    aria-label="{{ ::'kbn.docTable.tableRow.toggleRowDetailsButtonAriaLabel' | i18n: {defaultMessage: 'Toggle row details'} }}"
   >
     <icon ng-if="open" type="'arrowDown'" size="'s'"></icon>
     <icon ng-if="!open" type="'arrowRight'" size="'s'"></icon>
