+++ src/legacy/core_plugins/kibana/public/discover/doc_table/doc_table.html
@@ -113,7 +113,7 @@
     <div class="euiSpacer euiSpacer--m"></div>
 
     <p
-      i18n-id="common.ui.docTable.noResultsTitle"
+      i18n-id="kbn.docTable.noResultsTitle"
       i18n-default-message="No results found"
     >
     </p>
