+++ src/legacy/core_plugins/kibana/public/discover/doc_table/doc_table.js
@@ -23,13 +23,15 @@ import { getSort } from './lib/get_sort';
 import './infinite_scroll';
 import './components/table_header';
 import './components/table_row';
-import { dispatchRenderComplete } from '../render_complete';
-import { uiModules } from '../modules';
-import { getRequestInspectorStats, getResponseInspectorStats } from '../courier/utils/courier_inspector_utils';
+import { dispatchRenderComplete } from 'ui/render_complete';
+import { uiModules } from 'ui/modules';
+import 'ui/pager_control';
+import 'ui/pager';
+import { getRequestInspectorStats, getResponseInspectorStats } from 'ui/courier/utils/courier_inspector_utils';
 
 import { getLimitedSearchResultsMessage } from './doc_table_strings';
 
-uiModules.get('kibana')
+uiModules.get('app/discover')
   .directive('docTable', function (config, Notifier, getAppState, pagerFactory, $filter, courier, i18n) {
     return {
       restrict: 'E',
@@ -135,10 +137,10 @@ uiModules.get('kibana')
             let inspectorRequest = undefined;
             if (_.has($scope, 'inspectorAdapters.requests')) {
               $scope.inspectorAdapters.requests.reset();
-              const title = i18n('common.ui.docTable.inspectorRequestDataTitle', {
+              const title = i18n('kbn.docTable.inspectorRequestDataTitle', {
                 defaultMessage: 'Data',
               });
-              const description = i18n('common.ui.docTable.inspectorRequestDescription', {
+              const description = i18n('kbn.docTable.inspectorRequestDescription', {
                 defaultMessage: 'This request queries Elasticsearch to fetch the data for the search.',
               });
               inspectorRequest = $scope.inspectorAdapters.requests.start(title, { description });
