+++ src/legacy/core_plugins/kibana/public/discover/doc_table/doc_table_strings.js
@@ -25,7 +25,7 @@ import { i18n } from '@kbn/i18n';
  * @param resultCount {Number}
  */
 export function getLimitedSearchResultsMessage(resultCount) {
-  return i18n.translate('common.ui.docTable.limitedSearchResultLabel',
+  return i18n.translate('kbn.docTable.limitedSearchResultLabel',
     { defaultMessage: 'Limited to {resultCount} results. Refine your search.', values: { resultCount } }
   );
 }
