+++ src/legacy/core_plugins/kibana/public/discover/doc_table/infinite_scroll.js
@@ -18,7 +18,7 @@
  */
 
 import $ from 'jquery';
-import { uiModules } from '../modules';
+import { uiModules } from 'ui/modules';
 const module = uiModules.get('app/discover');
 
 module.directive('kbnInfiniteScroll', function () {
