+++ src/legacy/core_plugins/kibana/public/discover/doc_table/lib/get_sort.js
