+++ src/legacy/core_plugins/kibana/public/discover/embeddable/search_embeddable.ts
@@ -20,7 +20,6 @@
 import angular from 'angular';
 import _ from 'lodash';
 import { SearchSource } from 'ui/courier';
-import * as columnActions from 'ui/doc_table/actions/columns';
 import {
   ContainerState,
   Embeddable,
@@ -32,6 +31,7 @@ import { Filters, Query } from 'ui/embeddable/types';
 import { RequestAdapter } from 'ui/inspector/adapters';
 import { Adapters } from 'ui/inspector/types';
 import { getTime } from 'ui/timefilter/get_time';
+import * as columnActions from '../doc_table/actions/columns';
 import { SavedSearch } from '../types';
 import searchTemplate from './search_template.html';
