+++ src/legacy/core_plugins/kibana/public/discover/embeddable/search_embeddable_factory.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import 'ui/doc_table';
+import '../doc_table';
 
 import { i18n } from '@kbn/i18n';
 import { EmbeddableFactory } from 'ui/embeddable';
