+++ src/legacy/core_plugins/kibana/public/discover/index.js
@@ -22,7 +22,7 @@ import './directives';
 import 'ui/collapsible_sidebar';
 import './components/field_chooser/field_chooser';
 import './controllers/discover';
-import 'ui/doc_table/components/table_row';
+import './doc_table/components/table_row';
 import { FeatureCatalogueRegistryProvider, FeatureCatalogueCategory } from 'ui/registry/feature_catalogue';
 
 FeatureCatalogueRegistryProvider.register(i18n => {
