+++ src/legacy/core_plugins/kibana/public/doc_viewer/_doc_viewer.scss
