+++ src/legacy/core_plugins/kibana/public/doc_viewer/index.js
