+++ src/legacy/core_plugins/kibana/public/index.scss
@@ -24,6 +24,9 @@
 // Management styles
 @import './management/index';
 
+// Doc Viewer
+@import './doc_viewer/index';
+
 // Dashboard styles
 // MUST STAY AT THE BOTTOM BECAUSE OF DARK THEME IMPORTS
 @import './dashboard/index';
