+++ src/legacy/ui/public/_index.scss
@@ -12,8 +12,6 @@
 @import './chrome/index';
 @import './collapsible_sidebar/index';
 @import './directives/index';
-@import './doc_table/index';
-@import './doc_viewer/index';
 @import './error_allow_explicit_index/index';
 @import './error_auto_create_index/index';
 @import './error_url_overflow/index';
@@ -24,7 +22,6 @@
 // @import './timepicker/index';
 @import './markdown/index';
 @import './notify/index';
-@import './partials/index';
 @import './query_bar/index';
 @import './share/index';
 @import './filter_bar/index';
