+++ src/legacy/ui/public/directives/_index.scss
@@ -1,1 +1,2 @@
 @import './input_datetime';
+@import './saved_object_paginated_selectable'
