+++ src/legacy/ui/public/directives/_saved_object_paginated_selectable.scss
