+++ src/legacy/ui/public/directives/paginated_selectable_list.js
@@ -21,7 +21,7 @@ import _ from 'lodash';
 import { uiModules } from '../modules';
 import './paginate';
 import './kbn_href';
-import paginatedSelectableListTemplate from '../partials/paginated_selectable_list.html';
+import paginatedSelectableListTemplate from './partials/paginated_selectable_list.html';
 
 const module = uiModules.get('kibana');
