+++ src/legacy/ui/public/directives/partials/paginate_controls.html
