+++ src/legacy/ui/public/directives/partials/paginated_selectable_list.html
