+++ src/legacy/ui/public/directives/partials/saved_object_finder.html
