+++ src/legacy/ui/public/directives/saved_object_finder.js
@@ -23,7 +23,7 @@ import { keyMap } from '../utils/key_map';
 import { SavedObjectRegistryProvider } from '../saved_objects/saved_object_registry';
 import { uiModules } from '../modules';
 import '../filters/label';
-import savedObjectFinderTemplate from '../partials/saved_object_finder.html';
+import savedObjectFinderTemplate from './partials/saved_object_finder.html';
 import './input_focus';
 import './paginate';
