+++ x-pack/plugins/dashboard_mode/public/dashboard_viewer.js
@@ -36,8 +36,6 @@ import 'ui/vislib';
 import 'ui/agg_response';
 import 'ui/agg_types';
 import 'ui/timepicker';
-import 'ui/pager';
-import 'ui/pager_control';
 import 'leaflet';
 
 import { showAppRedirectNotification } from 'ui/notify';
