+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -299,28 +299,6 @@
     "common.ui.directives.paginate.size.allDropDownOptionLabel": "全部",
     "common.ui.directives.truncated.showLessLinkText": "更少",
     "common.ui.directives.truncated.showMoreLinkText": "更多",
-    "common.ui.docTable.inspectorRequestDataTitle": "数据",
-    "common.ui.docTable.inspectorRequestDescription": "此请求将查询 Elasticsearch 以获取搜索的数据。",
-    "common.ui.docTable.limitedSearchResultLabel": "仅限 {resultCount} 个结果。优化您的搜索。",
-    "common.ui.docTable.noResultsTitle": "找不到结果",
-    "common.ui.docTable.tableHeader.moveColumnLeftButtonAriaLabel": "向左移动“{columnName}”列",
-    "common.ui.docTable.tableHeader.moveColumnLeftButtonTooltip": "向左移动列",
-    "common.ui.docTable.tableHeader.moveColumnRightButtonAriaLabel": "向右移动“{columnName}”列",
-    "common.ui.docTable.tableHeader.moveColumnRightButtonTooltip": "向右移动列",
-    "common.ui.docTable.tableHeader.removeColumnButtonAriaLabel": "删除“{columnName}”列",
-    "common.ui.docTable.tableHeader.removeColumnButtonTooltip": "删除列",
-    "common.ui.docTable.tableHeader.sortByColumnAscendingAriaLabel": "升序排序 {columnName}",
-    "common.ui.docTable.tableHeader.sortByColumnDescendingAriaLabel": "降序排序 {columnName}",
-    "common.ui.docTable.tableHeader.sortByColumnTooltip": "按“{columnName}”排序",
-    "common.ui.docTable.tableHeader.sortByTimeTooltip": "按时间排序",
-    "common.ui.docTable.tableHeader.timeHeaderCellTitle": "时间",
-    "common.ui.docTable.tableRow.filterForValueButtonAriaLabel": "筛留值",
-    "common.ui.docTable.tableRow.filterForValueButtonTooltip": "筛留值",
-    "common.ui.docTable.tableRow.filterOutValueButtonAriaLabel": "筛除值",
-    "common.ui.docTable.tableRow.filterOutValueButtonTooltip": "筛除值",
-    "common.ui.docTable.tableRow.toggleRowDetailsButtonAriaLabel": "切换行详细信息",
-    "common.ui.docTable.tableRow.viewSingleDocumentLinkText": "查看单个文档",
-    "common.ui.docTable.tableRow.viewSurroundingDocumentsLinkText": "查看周围文档",
     "common.ui.errorAllowExplicitIndex.errorDescription": "似乎您的 Elasticsearch 集群已将设置 {allowExplicitIndexConfig} 设置为 {allowExplicitIndexValue}，这使 Kibana 无法执行搜索请求。使用此功能，我们可以向 Elasticsearch 发送单个请求来搜索多个索引，这样，当仪表板上有多个面板时，面板可快速且一致地加载。",
     "common.ui.errorAllowExplicitIndex.errorDisclaimer": "但是，只有解决了此问题后，您才能使用 Kibana 中的某些应用，如 Discover、Visualize 和仪表板。",
     "common.ui.errorAllowExplicitIndex.errorTitle": "糟糕！",
@@ -1310,6 +1288,28 @@
     "kbn.doc.failedToLocateDocumentDescription": "无法找到文档",
     "kbn.doc.loadingDescription": "正在加载……",
     "kbn.doc.somethingWentWrongDescription": "哎呦。出了问题。不是我找不到您的文档，而是无法尝试。索引缺失或类型缺失。去问问 Elasticsearch，似乎哪里不对劲。",
+    "kbn.docTable.inspectorRequestDataTitle": "数据",
+    "kbn.docTable.inspectorRequestDescription": "此请求将查询 Elasticsearch 以获取搜索的数据。",
+    "kbn.docTable.limitedSearchResultLabel": "仅限 {resultCount} 个结果。优化您的搜索。",
+    "kbn.docTable.noResultsTitle": "找不到结果",
+    "kbn.docTable.tableHeader.moveColumnLeftButtonAriaLabel": "向左移动“{columnName}”列",
+    "kbn.docTable.tableHeader.moveColumnLeftButtonTooltip": "向左移动列",
+    "kbn.docTable.tableHeader.moveColumnRightButtonAriaLabel": "向右移动“{columnName}”列",
+    "kbn.docTable.tableHeader.moveColumnRightButtonTooltip": "向右移动列",
+    "kbn.docTable.tableHeader.removeColumnButtonAriaLabel": "删除“{columnName}”列",
+    "kbn.docTable.tableHeader.removeColumnButtonTooltip": "删除列",
+    "kbn.docTable.tableHeader.sortByColumnAscendingAriaLabel": "升序排序 {columnName}",
+    "kbn.docTable.tableHeader.sortByColumnDescendingAriaLabel": "降序排序 {columnName}",
+    "kbn.docTable.tableHeader.sortByColumnTooltip": "按“{columnName}”排序",
+    "kbn.docTable.tableHeader.sortByTimeTooltip": "按时间排序",
+    "kbn.docTable.tableHeader.timeHeaderCellTitle": "时间",
+    "kbn.docTable.tableRow.filterForValueButtonAriaLabel": "筛留值",
+    "kbn.docTable.tableRow.filterForValueButtonTooltip": "筛留值",
+    "kbn.docTable.tableRow.filterOutValueButtonAriaLabel": "筛除值",
+    "kbn.docTable.tableRow.filterOutValueButtonTooltip": "筛除值",
+    "kbn.docTable.tableRow.toggleRowDetailsButtonAriaLabel": "切换行详细信息",
+    "kbn.docTable.tableRow.viewSingleDocumentLinkText": "查看单个文档",
+    "kbn.docTable.tableRow.viewSurroundingDocumentsLinkText": "查看周围文档",
     "kbn.home.addData.addDataToKibanaDescription": "使用这些解决方案可快速将您的数据转换成预构建的仪表板和监测系统。",
     "kbn.home.addData.addDataToKibanaTitle": "将数据添加到 Kibana",
     "kbn.home.addData.apm.addApmButtonLabel": "添加 APM",
